requirements.txt
Django==4.2.11
django-cors-headers==3.10.0
djangorestframework==3.14.0
drf-yasg==1.21.4
psycopg2-binary==2.9.5
psycopg2==2.9.5
PyJWT==2.4.0